Improved separation of logic for handlers
