+++ src/debugger/cdp-proxy/CDPMessageHandlers/products/safariCDPMessageHandler.ts
